+++ lib/home_screen.dart
@@ -1,10 +1,13 @@
 import 'dart:typed_data';
 
 import 'package:dot_document/dot_document.dart';
-import 'package:dot_flutter_sdk_example/page_routes.dart';
+import 'package:dot_face_lite/dot_face_lite.dart';
 import 'package:flutter/material.dart';
 
-import 'document_auto_capture_screen.dart';
+import 'document_auto_capture/document_auto_capture_screen.dart';
+import 'face_auto_capture/face_auto_capture_screen.dart';
+import 'magnifeye_liveness/magnifeye_liveness_screen.dart';
+import 'page_routes.dart';
 
 class HomeScreen extends StatefulWidget {
   const HomeScreen({super.key});
@@ -33,7 +36,7 @@ class _HomeScreenState extends State<HomeScreen> {
         await DefaultAssetBundle.of(context).load('assets/dot_license.lic');
     await DotSdk.instance.initialize(DotSdkConfiguration(
         licenseBytes: licenseByteData.buffer.asUint8List(),
-        libraries: [DotDocumentLibrary()]));
+        libraries: [DotDocumentLibrary(), DotFaceLiteLibrary()]));
   }
 
   @override
@@ -77,6 +80,20 @@ class _HomeScreenState extends State<HomeScreen> {
             onPressed: () {
               Navigator.of(context).push(
                   createRouteWithoutAnimation(DocumentAutoCaptureScreen()));
+            }),
+        _createComponent(
+            title: 'Face Auto Capture',
+            subtitle: 'Basic component sample.',
+            onPressed: () {
+              Navigator.of(context)
+                  .push(createRouteWithoutAnimation(FaceAutoCaptureScreen()));
+            }),
+        _createComponent(
+            title: 'MagnifEye Liveness',
+            subtitle: 'Basic component sample.',
+            onPressed: () {
+              Navigator.of(context)
+                  .push(createRouteWithoutAnimation(MagnifEyeLivenessScreen()));
             }),
       ],
     );