+++ lib/magnifeye_liveness/conversion_extensions.dart
@@ -1,0 +1,23 @@
+import 'package:dot_face_lite/dot_face_lite.dart';
+
+import '../face/conversion_extensions.dart';
+
+extension MagnifEyeLivenessResultExt on MagnifEyeLivenessResult {
+  Map<String, dynamic> toJson() {
+    return {
+      'bgraRawImage': {
+        'bytes': '${bgraRawImage.bytes.length} bytes',
+        'size': {
+          'width': bgraRawImage.size.width,
+          'height': bgraRawImage.size.height,
+        },
+      },
+      'faceDetectorResult': {
+        'confidence': faceDetectorResult.confidence,
+        'normalizedRectangle': faceDetectorResult.normalizedRectangle?.toJson(),
+        'imageParameters': faceDetectorResult.imageParameters?.toJson(),
+      },
+      'content': '${content.length} bytes',
+    };
+  }
+}