+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 0.2.0 - 2024-02-16
+### Added
+- Face Auto Capture component sample.
+- MagnifEye Liveness component sample.
+
+### Changed
+- Update DOT SDK to 0.2.1.
+
 ## 0.1.0 - 2024-01-24
 ### Added
-- Document auto capture component example.
+- Document Auto Capture component sample.